+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
 
+- Add class for parsing Article response ([#13](https://github.com/andyw8/pocket-ruby/pull/13))
 - Fix Pocket::Client::Retrieve#retrieve params default ([#12](https://github.com/andyw8/pocket-ruby/pull/12))
 
 ## [0.0.8] - 2021-03-28